requirements.txt
absl-py==2.1.0
accelerate==0.32.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.4.0
ase==3.23.0
astunparse==1.6.3
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
binpacking==1.5.2
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.2
colorama==0.4.6
constants==0.6.0
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dacite==1.8.1
datasets==2.20.0
deepspeed==0.14.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
einops==0.8.0
eofs==1.4.1
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gguf==0.9.1
gmpy2==2.2.1
google-pasta==0.2.0
gradio==4.42.0
gradio_client==1.3.0
grpcio==1.66.1
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
interegular==0.3.3
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.5
lark==1.2.2
libclang==18.1.1
lit==18.1.8
llamafactory==0.8.4.dev0
llvmlite==0.43.0
lm-format-enforcer==0.10.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.3.4
ml-dtypes==0.4.1
modelscope==1.18.0
mpmath==1.3.0
ms-swift==2.5.0.dev0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
netCDF4==1.7.1.post2
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.43.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
oss2==2.19.0
outlines==0.0.46
packaging==24.1
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
peft==0.12.0
pillow==10.4.0
pip==24.2
prometheus_client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
protobuf==4.25.5
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.20.0
ruff==0.6.3
safetensors==0.4.4
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==69.5.1
shapely==2.0.6
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==2.1.3
starlette==0.38.2
statistics==1.0.3.5
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.43.4
transformers-stream-generator==0.0.5
triton==3.0.0
trl==0.9.6
typer==0.12.5
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vllm==0.6.0
vllm-flash-attn==2.6.1
watchfiles==0.24.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.43.0
wolframclient==1.4.0
wrapt==1.16.0
xarray==2024.7.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.6
zipp==3.20.1
